Testing notifications on delete from collection

+++ public/components/collection-view/collection-view.js
@@ -1,4 +1,4 @@
-define(['bridget', 'knockout', 'text!./collection-view.html', 'masonry', 'imagesloaded', 'app'], function (bridget, ko, template, masonry, imagesLoaded, app) {
+define(['bridget', 'knockout', 'text!./collection-view.html', 'masonry', 'imagesloaded', 'app', 'smoke'], function (bridget, ko, template, masonry, imagesLoaded, app) {
 
 	$.bridget('masonry', masonry);
 	var transDuration = '0.4s';
@@ -7,9 +7,7 @@ define(['bridget', 'knockout', 'text!./collection-view.html', 'masonry', 'images
 		transDuration = 0;
 	}
 
-
 	ko.bindingHandlers.scroll = {
-
 		updating: true,
 
 		init: function (element, valueAccessor, allBindingsAccessor) {
@@ -62,7 +60,6 @@ define(['bridget', 'knockout', 'text!./collection-view.html', 'masonry', 'images
 			});
 		}
 	};
-
 
 	ko.showMoreLess = function (initialText) {
 
@@ -197,11 +194,7 @@ define(['bridget', 'knockout', 'text!./collection-view.html', 'masonry', 'images
 				},
 				error: function (xhr, textStatus, errorThrown) {
 					self.loading(false);
-
-					$("#myModal").find("h4").html("An error occured");
-					$("#myModal").find("div.modal-body").html(errorThrown);
-
-					$("#myModal").modal('show');
+					$.smkAlert({text:'An error has occured', type:'danger', permanent: true});
 				}
 			});
 		};
@@ -278,42 +271,29 @@ define(['bridget', 'knockout', 'text!./collection-view.html', 'masonry', 'images
 		};
 
 		self.removeRecord = function (e) {
-			$("#myModal").find("h4").html("Delete item");
-			$("#myModal").find("div.modal-body").html("Are you sure you want to proceed?");
-			var footer = $("#myModal").find("div.modal-footer");
-			if (footer.is(':empty')) {
-				$("#myModal").find("div.modal-footer").append('<button type="button" class="btn btn-default" data-dismiss="modal">Cancel</button><a class="btn btn-danger btn-ok">Delete</a>');
-			}
-			$("#myModal").modal('show');
-			$('.btn-danger').on('click', function (event) {
-				$("#myModal").find("div.modal-footer").html('');
-				$("#myModal").remove("div.modal-footer");
-
-				var jsondata = JSON.stringify({
-					recId: e
-				});
-				$.ajax({
-					url: '/collection/' + self.id() + '/removeRecord?recId=' + e,
-					type: 'DELETE',
-					contentType: "application/json",
-					data: jsondata,
-					success: function (data, textStatus, xhr) {
-						self.citems.remove(e);
-						if ($("#" + e)) {
-							$("#" + e).remove();
+			$.smkConfirm({text:'Are you sure you want to permanently remove this item?', accept: 'Delete', cancel: 'Cancel'}, function (ee) {
+				if (ee) {
+					$.ajax({
+						url: '/collection/' + self.id() + '/removeRecord?recId=' + e,
+						type: 'DELETE',
+						contentType: "application/json",
+						data: JSON.stringify(e),
+						success: function (data, textStatus, xhr) {
+							self.citems.remove(e);
+							if ($("#" + e)) {
+								$("#" + e).remove();
+							}
+							self.itemCount(self.itemCount() - 1);
+							$.smkAlert({text:'Item removed from the collection', type:'success'});
+							$container.masonry( 'layout');
+						},
+						error: function (xhr, textStatus, errorThrown) {
+							$.smkAlert({text:'An error has occured', type:'danger', time: 10});
 						}
-						self.itemCount(self.itemCount() - 1);
-						$("#myModal").find("h4").html("Done!");
-						$("#myModal").find("div.modal-body").html("Item removed from collection");
-						$("#myModal").modal('show');
-						$container.masonry( 'layout');
-					},
-					error: function (xhr, textStatus, errorThrown) {
-						$("#myModal").find("h4").html("An error occured");
-						$("#myModal").find("div.modal-body").html(errorThrown);
-						$("#myModal").modal('show');
-					}
-				});
+					});
+				} else {
+
+				}
 			});
 		};
